Added very basic support for displaying a GIFs read from Unix.
Unfortunately the Unix/C file semantics and the Arduino FS ones
are quite different, so I didn't add unix support to all the code,
as reading directories, searching for files by index, and so forth
would be lots of ifdefs.
I only added the most basic necessary support in SimpleGifAnimViewer2.ino
which will load a single filename, but that's good enough for my needs.